change: remove XML and DOCTYPE
The usage of XML version and DOCTYPE is more for stand alone files,
i.e. it is one of the possible output adapter. In order to not restrict
the final SVG result and keep it more independent - this was removed.

Additional output adapter can be used to add that info.