update: position mixins below the node
Before a "container" for mixins was vertically centered.
After it's was aligned to the top of the node, so first mixed in node will be on the same Y-position as node of the interest. All other mixins will be positioned below.

This is required for creating small groups of nodes with its own children and mixins to allow nesting of that groups to create a bigger tree.