fix: revert back to SVG width/height
Reverting back the switching of width/height to viewBox. With viewBox the whole diagram is scaling to fit the page width. In this case the small diagram will be huge relative to the big diagram - it doesn't looks nice. Also when the screen is smaller than the width of the diagram, then the whole diagram is scaling down, so big diagrams can became useless.
It is better to have the same sizes for all diagrams, so any size of diagram will have the same size of boxes. Also it will be possible to zoom in and to navigate throw the diagram by using scrollbars.